Initializate Firebase Notification Service on app boot
Once the app has been launched, we will load firebase and the
notification service.

With this, notifications will have full access to the app, so it is not
page dependant.

If we are inside the application it will bring us a toastcontroller,
will send a notification if the app is on background or not opened.

Signed-off-by: Hernán Castañón <hernanderes34@gmail.com>